studied linked list, pointer and recursion

+++ Recursion/power.cpp
@@ -1,12 +1,12 @@
 #include<iostream>
 using namespace std;
 
-int findPower(int num, int power){
+long long findPower(int num, int power){
     if(power==0)
         return 1;
 
-    int a = findPower(num, power-1);
-    int b= num*a;
+    long long a = findPower(num, power-1);
+    long long b= num*a;
     return b;
 }
 
@@ -15,7 +15,7 @@ int main(){
     cout<<"Enter the number and power : ";
     cin>>num>>power;
 
-    int ans = findPower(num, power);
+    long long ans = findPower(num, power);
     cout<<ans<<endl; 
     return 0;
 }
